assembler: split expanded.rs into appropriate modules
